+++ R/gg_crossbar.R
@@ -82,7 +82,7 @@ gg_crossbar <- function(
     col_limits = NULL,
     col_oob = scales::oob_keep,
     col_palette = NULL,
-    col_palette_na = "#c8d7df",
+    col_palette_na = "#4b636e",
     col_rescale = scales::rescale(),
     col_steps = FALSE,
     col_title = NULL,